Se implementaron productos Base para que al iniciar, hayan algunos productos en la base de datos, también se agregaron imagenes en la página principal

+++ storechainparent/microservices/inventory/src/main/java/com/storechain/inventory/service/InventoryService.java
@@ -46,7 +46,7 @@ public class InventoryService {
                         HttpStatus.NOT_FOUND,
                         "Producto no encontrado"));
 
-        int newStock = product.getStock() + quantity;
+        Double newStock = product.getStock() + quantity;
 
         if (newStock < 0) {
             throw new BusinessRuleException("1004",
